requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-ollama
PyMuPDF
